Fixed a wrong cast double -> int
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@3449 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/org/objectweb/proactive/examples/nbody/barneshut/Domain.java
@@ -122,9 +122,8 @@ public class Domain implements Serializable {
                 logger.info("Compute movement.");
             }
         } else {
-            this.display.drawBody((int) this.rock.x, (int) this.rock.y,
-                (int) this.rock.z, (int) this.rock.vx, (int) this.rock.vy,
-                (int) this.rock.vz, (int) this.rock.mass,
+            this.display.drawBody(this.rock.x, this.rock.y, this.rock.z,
+                this.rock.vx, this.rock.vy, this.rock.vz, (int) this.rock.mass,
                 (int) this.rock.diameter, this.identification, this.hostName);
         }
     }
